Update 5-flip_bits.c - Fixed the type.

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,7 +12,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0, diff;
+	unsigned long int count = 0, diff;
 
 	diff = n ^ m;
